Move test files to notifications folder
